CSS additions
Cascading Style Sheets (CSS) are responsible for making the website look pretty (i.e. page alignment, background and texts colors, etc.) In one of the assignments this week, I was tasked with adding style to the index and boards pages. To do so, I created a .css file and linked it to the pages so the styling is applied. One difference to note is that applying header styles can vary by which heading you choose. For this project, I prefer only styling the main header (h1).