Week 2 assignment - practice with functions
To get a better understanding of how functions work, I am tasked with creating a game. The game I chose to code is Rock, Paper, Scissors a basic yet great example of how functions work and add to the UI. The functions used and created are play(userChoice), where the user chooses rock, paper, or scissors to initiate the game; a function for the computer to choose rock, paper, or scissors; and a function deciding the winner based on different combinations and outcomes.