Adding sound to Rock, Paper, Scissors game
I added an `introSound` for theme music, sounds for making each selection, and sounds for win, lose, or tie. I ran into an issue where the theme music wouldn't auto play in the [Chrome] browser, so I had to use functions to work around it. I did this without actionListeners.