CSC102 Assignment 7.1 Table Creation
Objectives:

- Table creation
- Use CSS to make your web application look and feel professional.
- Understand Inline vs Internal vs External CSS

Requirements
Create a table on a web page with a minimum of 3 columns and 4 rows of data.  This table should hold useful information, not be just a "sample table".

NOTE:
You MUST NOT use actionListeners
Remove ALL unused and example code (not commented out)
You CANNOT have javascript functions in HTML

I went back to a previous project and added a table with 15 ways to say "hello" in various languages. Instead of using a separate CSS (External) file, my CSS / design code is in the HTML file under the <style> tag. This is Internal CSS. I did not use any Inline CSS styling in this file, but have in other projects (see file index.html in the <img id> tag). Since I could not have JS functions in the HTML file, I had to create a separate JavaScript file for the sayHello function, since I still wanted it with the webpage.